+++ src/main/java/ex1/Entreprise.java
@@ -1,0 +1,21 @@
+package ex1;
+
+import java.util.Date;
+/** Entreprise */
+public class Entreprise {
+    /** Numéro de siret */
+    public int siret;
+    /** Nom */
+    public String nom;
+    /** Adresse */
+    public String adresse;
+    /** Date de création */
+    public Date dateCreation;
+    /** Capital maximal */
+    public static final int capitalMax = 3000000;
+
+    /**
+     * Afficher le status
+     */
+    public void afficherStatut(){}
+}
