--- src/main/java/ex1/entreprise.java
@@ -1,18 +1,0 @@
-package ex1;
-
-import java.util.Date;
-
-public class entreprise {
-
-	public int Siret;
-	public String Nom;
-	public String adresse;
-	public Date date_Creation;
-	
-	public static final int capitalMax = 3000000;
-	
-	public void Afficher_statut(){
-		
-	}
-	
-}
