--- src/main/java/ex2/LivretA.java
@@ -1,9 +1,0 @@
-package ex2;
-
-public class LivretA extends CompteBancaire {
-
-	public LivretA(String type, double solde, double tauxRemuneration) {
-		super(type, solde, 0, tauxRemuneration);
-	}
-
-}
