+++ src/main/java/ex2/LivretARate.java
@@ -1,0 +1,14 @@
+package ex2;
+/** Représente un compte bancaire de type livret A (type=LA)
+ * @author DIGINAMIC
+ */
+public class LivretARate extends CompteBancaireRate {
+    public LivretARate(String type,
+                       double solde,
+                       double tauxRemuneration) {
+        super(type,
+                solde,
+                0,
+                tauxRemuneration);
+    }
+}
